Checkout

+++ Ecommerce_Canecaria_DevGo/src/main/java/ecommerce/Model/MetodosAcessores/FinalizarCompra.java
@@ -20,7 +20,7 @@ public class FinalizarCompra {
     private String cidade;
     private String estado;
     private String codFinalizacaoCompra;
-    private float valorTotalCompra;
+    private double valorTotalCompra;
     private Timestamp dataPedido;
 
     
@@ -208,7 +208,7 @@ public class FinalizarCompra {
     /**
      * @return the valorTotalCompra
      */
-    public float getValorTotalCompra() {
+    public double getValorTotalCompra() {
         return valorTotalCompra;
     }
